Fix(metadata): Improve handling of non-author creator roles (#1987)
* Fix(metadata): Don't import creators with a non-author role as author

Looks at either the meta tag for the creators or the opf role attribute
to determine the role of a creator. If no role is specified its assumed
to be an author.

Also only prunes handled created roles when writing the epub data back
out to prevent clobbering any extra data that might be present in the
epub already.

* Fix(metadata): Add testcases for non-author creator handling

Just adds testcases for the previous changes to the extraction and
writer logic.